Split architecture policy tests
