+++ tests/unit/tooling/test_architecture_unsupported_surfaces.py
@@ -1,0 +1,121 @@
+"""Architecture checker contracts for unsupported-surface scanning."""
+
+from __future__ import annotations
+
+from pathlib import Path
+
+from tools.check_architecture import check_architecture
+
+
+def _write(root: Path, relative: str, source: str) -> Path:
+    path = root / relative
+    path.parent.mkdir(parents=True, exist_ok=True)
+    path.write_text(source, encoding="utf-8")
+    return path
+
+
+_K = "knowledge"
+_S = "search"
+_RM = "Research" + "Memory"
+_RE = "Research" + "Episode"
+_KS_DOT = _K + "." + _S
+
+
+def test_research_memory_import_in_src_is_flagged(tmp_path: Path) -> None:
+    _write(
+        tmp_path,
+        "src/jacobian/capability_service.py",
+        f"from jacobian.memory import {_RM}\n",
+    )
+    report = check_architecture(tmp_path)
+    surf = [v for v in report.violations if v.code == "unsupported-surface"]
+    assert len(surf) >= 1
+
+
+def test_research_memory_import_in_tests_is_flagged(tmp_path: Path) -> None:
+    """Unsupported surfaces are scanned in tests too, not just src."""
+    _write(
+        tmp_path,
+        "tests/composition/test_memory.py",
+        f"from jacobian.memory import {_RM}\n",
+    )
+    report = check_architecture(tmp_path)
+    surf = [v for v in report.violations if v.code == "unsupported-surface"]
+    assert len(surf) >= 1
+    assert surf[0].path == "tests/composition/test_memory.py"
+
+
+def test_knowledge_search_string_in_src_is_flagged(tmp_path: Path) -> None:
+    _write(
+        tmp_path,
+        "src/jacobian/builtin_capabilities.py",
+        f'capability_id = "{_KS_DOT}"\n',
+    )
+    report = check_architecture(tmp_path)
+    surf = [v for v in report.violations if v.code == "unsupported-surface"]
+    assert len(surf) == 1
+    assert _KS_DOT in surf[0].message
+
+
+def test_knowledge_search_in_docs_is_flagged(tmp_path: Path) -> None:
+    _write(
+        tmp_path,
+        "docs/explanation/search.md",
+        f"# Search\n\nUse {_KS_DOT} for retrieval.\n",
+    )
+    report = check_architecture(tmp_path)
+    surf = [v for v in report.violations if v.code == "unsupported-surface"]
+    assert len(surf) == 1
+    assert surf[0].path == "docs/explanation/search.md"
+
+
+def test_knowledge_search_in_schema_is_flagged(tmp_path: Path) -> None:
+    _write(
+        tmp_path,
+        "benchmarks/schemas/test-schema.json",
+        f'{{"capability_id": "{_KS_DOT}"}}\n',
+    )
+    report = check_architecture(tmp_path)
+    surf = [v for v in report.violations if v.code == "unsupported-surface"]
+    assert len(surf) == 1
+    assert surf[0].path == "benchmarks/schemas/test-schema.json"
+
+
+def test_research_episode_in_docs_is_flagged(tmp_path: Path) -> None:
+    _write(
+        tmp_path,
+        "docs/explanation/episodes.md",
+        f"# Episodes\n\nThe {_RE} store records past work.\n",
+    )
+    report = check_architecture(tmp_path)
+    surf = [v for v in report.violations if v.code == "unsupported-surface"]
+    assert len(surf) >= 1
+    assert surf[0].path == "docs/explanation/episodes.md"
+
+
+def test_research_memory_prose_in_docs_is_flagged(tmp_path: Path) -> None:
+    _write(
+        tmp_path,
+        "docs/explanation/memory.md",
+        "# Memory\n\nThe Research memory store is deprecated.\n",
+    )
+    report = check_architecture(tmp_path)
+    surf = [v for v in report.violations if v.code == "unsupported-surface"]
+    assert len(surf) >= 1
+
+
+def test_generic_memory_word_is_not_flagged(tmp_path: Path) -> None:
+    """Legitimate generic words like 'memory limits' must not trigger."""
+    _write(
+        tmp_path,
+        "src/jacobian/process_policy.py",
+        "import os\n\n# Configure memory limits for child processes.\n",
+    )
+    report = check_architecture(tmp_path)
+    assert all(v.code != "unsupported-surface" for v in report.violations)
+
+
+def test_changelog_is_excluded_from_surface_scan(tmp_path: Path) -> None:
+    _write(tmp_path, "CHANGELOG.md", f"# Changelog\n\nAdded {_KS_DOT} capability.\n")
+    report = check_architecture(tmp_path)
+    assert all(v.code != "unsupported-surface" for v in report.violations)
